Move era customizations to _cfi's
