+++ RecoPixelVertexing/PixelTriplets/python/pixelTripletHLTEDProducer_cfi.py
@@ -1,4 +1,6 @@
-from RecoPixelVertexing.PixelTriplets.pixelTripletHLTEDProducer_cfi import *
+from RecoPixelVertexing.PixelTriplets.pixelTripletHLTEDProducerDefault_cfi import pixelTripletHLTEDProducerDefault as _pixelTripletHLTEDProducerDefault
+
+pixelTripletHLTEDProducer = _pixelTripletHLTEDProducerDefault.clone()
 from Configuration.Eras.Modifier_trackingLowPU_cff import trackingLowPU
 from Configuration.Eras.Modifier_trackingPhase1PU70_cff import trackingPhase1PU70
 from Configuration.Eras.Modifier_trackingPhase2PU140_cff import trackingPhase2PU140
