+++ RecoTracker/TkHitPairs/python/hitPairEDProducer_cfi.py
@@ -1,4 +1,6 @@
-from RecoTracker.TkHitPairs.hitPairEDProducer_cfi import *
+from RecoTracker.TkHitPairs.hitPairEDProducerDefault_cfi import hitPairEDProducerDefault as _hitPairEDProducerDefault
+
+hitPairEDProducer = _hitPairEDProducerDefault.clone()
 from Configuration.Eras.Modifier_trackingPhase1PU70_cff import trackingPhase1PU70
 from Configuration.Eras.Modifier_trackingPhase2PU140_cff import trackingPhase2PU140
 trackingPhase1PU70.toModify(hitPairEDProducer, maxElement=0)
